# src/server/api/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SupportedDbIdentifier": {
				"enum": [
					"postgresql"
				],
				"type": "string"
			},
			"DatabasePluginConfig": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/SupportedDbIdentifier"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ConnectionConfig": {
				"properties": {
					"pluginId": {
						"$ref": "#/components/schemas/SupportedDbIdentifier"
					},
					"name": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "number",
						"format": "double"
					},
					"database": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"pluginId",
					"name",
					"host",
					"port",
					"database"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Connection": {
				"properties": {
					"connectionId": {
						"type": "string"
					},
					"connectionConfig": {
						"$ref": "#/components/schemas/ConnectionConfig"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"connectionId",
					"connectionConfig",
					"createdAt",
					"isActive"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ConnectionResult": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"connectionTime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "elytra",
		"version": "1.0.0",
		"description": "The most powerful enterprise AI-driven database client.",
		"contact": {}
	},
	"paths": {
		"/trigger/health": {
			"get": {
				"operationId": "Health",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/plugins/ids": {
			"get": {
				"operationId": "GetSupportedDbIds",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SupportedDbIdentifier"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/plugins/configs": {
			"get": {
				"operationId": "GetSupportedDbConfigs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DatabasePluginConfig"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/plugins/{id}/config": {
			"get": {
				"operationId": "GetSupportedDbConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabasePluginConfig"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportedDbIdentifier"
						}
					}
				]
			}
		},
		"/connections/all": {
			"get": {
				"operationId": "GetAllConnections",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Connection"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/connections/create": {
			"post": {
				"operationId": "CreateConnection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"$ref": "#/components/schemas/ConnectionResult"
										},
										"connection": {
											"$ref": "#/components/schemas/Connection"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConnectionConfig"
							}
						}
					}
				}
			}
		},
		"/connections/test": {
			"post": {
				"operationId": "TestConnection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConnectionResult"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConnectionConfig"
							}
						}
					}
				}
			}
		},
		"/connections/{connectionId}/close": {
			"post": {
				"operationId": "CloseConnection",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}